Separate DNN and HME into two payloads, change payloadProducers to only HME
